third_party/blink/web_tests/platform/win11-arm64/virtual/fenced-frame-mparch/external/wpt/html/anonymous-iframe/embedding.tentative.https.window_9-9-expected.txt

This is a testharness.js-based test.
FAIL COOP:same-origin + COEP:credentialless embeds same-origin credentialless iframe assert_equals: expected "load" but got "block"
Harness: the test ran to completion.
